config_lib/pyproject.toml
[tool.poetry]
name = "config-lib"
version = "0.1.0"
description = ""
authors = ["Alex-Evo-2266 <ghbdtn2266@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
pyyaml = "^6.0.2"
pytest = "^8.4.2"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
